Truly fixed #2. Changed way of getting distro.
The last commit solved the issue I was having with GalliumOs but it did
not really solve the problem of unknown systems. Now there will not be a
message of 'update complete' when a system is unidentified. Also, this
commit solves the bug of /etc/os-release 's that have quotation marks in
their ID's not being properly handled.