Kernel infomration printed. Minor commit
Added the name for the system (hostname) to be displayed while printing
information. Also added the ability to pass in argument of -s to only
show sysinfo and skip the update process. No longer display the desktop
session at beginning. This will solve the issue of GUI-less desktops
showing a blank for session.