Fixed #5.
Added a try / except for the error subprocess was raising when an incorrect password was entered. Added looping capabilities to allow a user to retry when entered an incorrect password. May possbily need to add some timeout function to determine if the user is attempting to be malicious (updating can't be too malicious can it?).