Slackware was added, currently testing.
Slackware capability was added, there are still a few things to iron out.
Right now, slackpkg is used and it is required to run the script as root,
rather than going to root for each of the commands. The IP address will not
show after an update, as the hostname command is slightly different on slackware.